+++ rlberry/agents/torch/a2c/a2c.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 
 import gymnasium.spaces as spaces
 import numpy as np
-from rlberry.agents import AgentWithSimplePolicy
+from rlberry.agents import AgentWithSimplePolicy, AgentTorch
 from rlberry.agents.utils.replay import ReplayBuffer
 from rlberry.agents.torch.utils.training import optimizer_factory
 from rlberry.agents.torch.utils.models import default_policy_net_fn
@@ -17,7 +17,7 @@ import rlberry
 logger = rlberry.logger
 
 
-class A2CAgent(AgentWithSimplePolicy):
+class A2CAgent(AgentTorch, AgentWithSimplePolicy):
     """
     Advantage Actor Critic Agent.