+++ rlberry/agents/__init__.py
@@ -14,3 +14,4 @@ from .optql import OptQLAgent
 from .psrl import PSRLAgent
 from .rlsvi import RLSVIAgent
 from .ucbvi import UCBVIAgent
+from .tabular_rl import QLAgent, SARSAAgent