+++ rlberry/agents/torch/reinforce/reinforce.py
@@ -1,5 +1,6 @@
 import torch
 import inspect
+import numpy as np
 
 import gymnasium.spaces as spaces
 from rlberry.agents import AgentWithSimplePolicy, AgentTorch
@@ -200,7 +201,7 @@ class REINFORCEAgent(AgentTorch, AgentWithSimplePolicy):
             rewards.insert(0, discounted_reward)
 
         # convert list to tensor
-        states = torch.FloatTensor(self.memory.states).to(self.device)
+        states = torch.FloatTensor(np.array(self.memory.states)).to(self.device)
         actions = torch.LongTensor(self.memory.actions).to(self.device)
         rewards = torch.FloatTensor(rewards).to(self.device)
         if self.normalize:
@@ -228,9 +229,9 @@ class REINFORCEAgent(AgentTorch, AgentWithSimplePolicy):
     def sample_parameters(cls, trial):
         batch_size = trial.suggest_categorical("batch_size", [1, 4, 8, 16, 32])
         gamma = trial.suggest_categorical("gamma", [0.9, 0.95, 0.99])
-        learning_rate = trial.suggest_loguniform("learning_rate", 1e-5, 1)
+        learning_rate = trial.suggest_float("learning_rate", 1e-5, 1, log=True)
 
-        entr_coef = trial.suggest_loguniform("entr_coef", 1e-8, 0.1)
+        entr_coef = trial.suggest_float("entr_coef", 1e-8, 0.1, log=True)
 
         return {
             "batch_size": batch_size,